@@ -20,14 +20,14 @@ namespace saba
 	MMDLookAtCamera::MMDLookAtCamera(const MMDCamera & cam)
 	{
 		glm::mat4 view(1.0f);
-		view = glm::translate(view, glm::vec3(0, 0, std::abs(cam.m_distance)));
+		view = translate(view, glm::vec3(0, 0, std::abs(cam.m_distance)));
 		//view = glm::mat4_cast(cam.m_rotate) * view;
 		//glm::mat4 rot;
-		auto degree = glm::degrees(cam.m_rotate);
+		//auto degree = degrees(cam.m_rotate);
 		glm::mat4 rot(1.0f);
-		rot = glm::rotate(rot, cam.m_rotate.y, glm::vec3(0, 1, 0));
-		rot = glm::rotate(rot, cam.m_rotate.z, glm::vec3(0, 0, -1));
-		rot = glm::rotate(rot, cam.m_rotate.x, glm::vec3(1, 0, 0));
+		rot = rotate(rot, cam.m_rotate.y, glm::vec3(0, 1, 0));
+		rot = rotate(rot, cam.m_rotate.z, glm::vec3(0, 0, -1));
+		rot = rotate(rot, cam.m_rotate.x, glm::vec3(1, 0, 0));
 		view = rot * view;
 
 		m_eye = glm::vec3(view[3]) + cam.m_interest;
