@@ -6,7 +6,6 @@
 #ifndef SABA_MODEL_MMD_MMDCAMERA_H_
 #define SABA_MODEL_MMD_MMDCAMERA_H_
 
-#include <glm/vec3.hpp>
 #include <glm/gtc/quaternion.hpp>
 
 namespace saba
@@ -15,19 +14,19 @@ namespace saba
 	{
 		MMDCamera();
 
-		glm::vec3	m_interest;
-		glm::vec3	m_rotate;
-		float		m_distance;
-		float		m_fov;
+		glm::vec3	m_interest{};
+		glm::vec3	m_rotate{};
+		float		m_distance{};
+		float		m_fov{};
 	};
 
 	struct MMDLookAtCamera
 	{
 		explicit MMDLookAtCamera(const MMDCamera& cam);
 
-		glm::vec3	m_center;
-		glm::vec3	m_eye;
-		glm::vec3	m_up;
+		glm::vec3	m_center{};
+		glm::vec3	m_eye{};
+		glm::vec3	m_up{};
 	};
 }
 
