@@ -18,7 +18,7 @@ namespace saba
 		void SetName(const std::string& name) { m_name = name; }
 		const std::string& GetName() const { return m_name; }
 
-		void SetWeight(float weight) { m_weight = weight; }
+		void SetWeight(const float weight) { m_weight = weight; }
 		float GetWeight() const { return m_weight; }
 
 		void SaveBaseAnimation() { m_saveAnimWeight = m_weight; }
