@@ -28,21 +28,21 @@ namespace saba
 				return false;
 			}
 
-			auto& header = pmdFile->m_header;
-			Read(&header.m_magic, file);
-			Read(&header.m_version, file);
-			Read(&header.m_modelName, file);
-			Read(&header.m_comment, file);
+			auto& [m_magic, m_version, m_modelName, m_comment, m_haveEnglishNameExt, m_englishModelNameExt, m_englishCommentExt] = pmdFile->m_header;
+			Read(&m_magic, file);
+			Read(&m_version, file);
+			Read(&m_modelName, file);
+			Read(&m_comment, file);
 
-			header.m_haveEnglishNameExt = 0;
+			m_haveEnglishNameExt = 0;
 
-			if (header.m_magic.ToString() != "Pmd")
+			if (m_magic.ToString() != "Pmd")
 			{
 				SABA_ERROR("PMD Header Error.");
 				return false;
 			}
 
-			if (header.m_version != 1.0f)
+			if (m_version != 1.0f)
 			{
 				SABA_ERROR("PMD Version Error.");
 				return false;
@@ -66,15 +66,15 @@ namespace saba
 
 			auto& vertices = pmdFile->m_vertices;
 			vertices.resize(vertexCount);
-			for (auto& vertex : vertices)
+			for (auto& [m_position, m_normal, m_uv, m_bone, m_boneWeight, m_edge] : vertices)
 			{
-				Read(&vertex.m_position, file);
-				Read(&vertex.m_normal, file);
-				Read(&vertex.m_uv, file);
-				Read(&vertex.m_bone[0], file);
-				Read(&vertex.m_bone[1], file);
-				Read(&vertex.m_boneWeight, file);
-				Read(&vertex.m_edge, file);
+				Read(&m_position, file);
+				Read(&m_normal, file);
+				Read(&m_uv, file);
+				Read(&m_bone[0], file);
+				Read(&m_bone[1], file);
+				Read(&m_boneWeight, file);
+				Read(&m_edge, file);
 			}
 
 			return !file.IsBad();
@@ -95,11 +95,11 @@ namespace saba
 
 			auto& faces = pmdFile->m_faces;
 			faces.resize(faceCount / 3);
-			for (auto& face : faces)
+			for (auto& [m_vertices] : faces)
 			{
-				Read(&face.m_vertices[0], file);
-				Read(&face.m_vertices[1], file);
-				Read(&face.m_vertices[2], file);
+				Read(&m_vertices[0], file);
+				Read(&m_vertices[1], file);
+				Read(&m_vertices[2], file);
 			}
 
 			return !file.IsBad();
@@ -120,17 +120,17 @@ namespace saba
 
 			auto& materials = pmdFile->m_materials;
 			materials.resize(materialCount);
-			for (auto& material : materials)
+			for (auto& [m_diffuse, m_alpha, m_specularPower, m_specular, m_ambient, m_toonIndex, m_edgeFlag, m_faceVertexCount, m_textureName] : materials)
 			{
-				Read(&material.m_diffuse, file);
-				Read(&material.m_alpha, file);
-				Read(&material.m_specularPower, file);
-				Read(&material.m_specular, file);
-				Read(&material.m_ambient, file);
-				Read(&material.m_toonIndex, file);
-				Read(&material.m_edgeFlag, file);
-				Read(&material.m_faceVertexCount, file);
-				Read(&material.m_textureName, file);
+				Read(&m_diffuse, file);
+				Read(&m_alpha, file);
+				Read(&m_specularPower, file);
+				Read(&m_specular, file);
+				Read(&m_ambient, file);
+				Read(&m_toonIndex, file);
+				Read(&m_edgeFlag, file);
+				Read(&m_faceVertexCount, file);
+				Read(&m_textureName, file);
 			}
 
 			return !file.IsBad();
@@ -151,14 +151,14 @@ namespace saba
 
 			auto& bones = pmdFile->m_bones;
 			bones.resize(boneCount);
-			for (auto& bone : bones)
+			for (auto& [m_boneName, m_parent, m_tail, m_boneType, m_ikParent, m_position, m_englishBoneNameExt] : bones)
 			{
-				Read(&bone.m_boneName, file);
-				Read(&bone.m_parent, file);
-				Read(&bone.m_tail, file);
-				Read(&bone.m_boneType, file);
-				Read(&bone.m_ikParent, file);
-				Read(&bone.m_position, file);
+				Read(&m_boneName, file);
+				Read(&m_parent, file);
+				Read(&m_tail, file);
+				Read(&m_boneType, file);
+				Read(&m_ikParent, file);
+				Read(&m_position, file);
 			}
 
 			return !file.IsBad();
@@ -179,16 +179,16 @@ namespace saba
 
 			auto& iks = pmdFile->m_iks;
 			iks.resize(ikCount);
-			for (auto& ik : iks)
+			for (auto& [m_ikNode, m_ikTarget, m_numChain, m_numIteration, m_rotateLimit, m_chanins] : iks)
 			{
-				Read(&ik.m_ikNode, file);
-				Read(&ik.m_ikTarget, file);
-				Read(&ik.m_numChain, file);
-				Read(&ik.m_numIteration, file);
-				Read(&ik.m_rotateLimit, file);
+				Read(&m_ikNode, file);
+				Read(&m_ikTarget, file);
+				Read(&m_numChain, file);
+				Read(&m_numIteration, file);
+				Read(&m_rotateLimit, file);
 
-				ik.m_chanins.resize(ik.m_numChain);
-				for (auto& ikChain : ik.m_chanins)
+				m_chanins.resize(m_numChain);
+				for (auto& ikChain : m_chanins)
 				{
 					Read(&ikChain, file);
 				}
@@ -212,21 +212,21 @@ namespace saba
 
 			auto& morphs = pmdFile->m_morphs;
 			morphs.resize(blendShapeCount);
-			for (auto& morph : morphs)
+			for (auto& [m_morphName, m_morphType, m_vertices, m_englishShapeNameExt] : morphs)
 			{
-				Read(&morph.m_morphName, file);
+				Read(&m_morphName, file);
 
 				uint32_t vertexCount = 0;
 				Read(&vertexCount, file);
-				morph.m_vertices.resize(vertexCount);
+				m_vertices.resize(vertexCount);
 
 				uint8_t morphType;
 				Read(&morphType, file);
-				morph.m_morphType = static_cast<PMDMorph::MorphType>(morphType);
-				for (auto& vtx : morph.m_vertices)
+				m_morphType = static_cast<PMDMorph::MorphType>(morphType);
+				for (auto& [m_vertexIndex, m_position] : m_vertices)
 				{
-					Read(&vtx.m_vertexIndex, file);
-					Read(&vtx.m_position, file);
+					Read(&m_vertexIndex, file);
+					Read(&m_position, file);
 				}
 			}
 
@@ -269,9 +269,9 @@ namespace saba
 			}
 
 			// ボーンの枠にはデフォルトでセンターが用意されているので、サイズを一つ多くする
-			pmdFile->m_boneDisplayLists.resize(size_t(displayListCount) + 1);
+			pmdFile->m_boneDisplayLists.resize(static_cast<size_t>(displayListCount) + 1);
 			bool first = true;
-			for (auto& displayList : pmdFile->m_boneDisplayLists)
+			for (auto& [m_name, m_displayList, m_englishNameExt] : pmdFile->m_boneDisplayLists)
 			{
 				if (first)
 				{
@@ -279,7 +279,7 @@ namespace saba
 				}
 				else
 				{
-					Read(&displayList.m_name, file);
+					Read(&m_name, file);
 				}
 			}
 
@@ -319,31 +319,31 @@ namespace saba
 				Read(&pmdFile->m_header.m_englishCommentExt, file);
 
 				// BoneName
-				for (auto& bone : pmdFile->m_bones)
+				for (auto& [m_boneName, m_parent, m_tail, m_boneType, m_ikParent, m_position, m_englishBoneNameExt] : pmdFile->m_bones)
 				{
-					Read(&bone.m_englishBoneNameExt, file);
+					Read(&m_englishBoneNameExt, file);
 				}
 
 				// BlendShape Name
 				/*
 				BlendShapeの最初はbaseなので、EnglishNameを読まないようにする
 				*/
-				size_t numBlensShape = pmdFile->m_morphs.size();
+				const size_t numBlensShape = pmdFile->m_morphs.size();
 				for (size_t bsIdx = 1; bsIdx < numBlensShape; bsIdx++)
 				{
-					auto& morpsh = pmdFile->m_morphs[bsIdx];
-					Read(&morpsh.m_englishShapeNameExt, file);
+					auto& [m_morphName, m_morphType, m_vertices, m_englishShapeNameExt] = pmdFile->m_morphs[bsIdx];
+					Read(&m_englishShapeNameExt, file);
 				}
 
 				// BoneDisplayNameLists
 				/*
 				BoneDisplayNameの最初はセンターとして使用しているので、EnglishNameを読まないようにする
 				*/
-				size_t numBoneDisplayName = pmdFile->m_boneDisplayLists.size();
+				const size_t numBoneDisplayName = pmdFile->m_boneDisplayLists.size();
 				for (size_t displayIdx = 1; displayIdx < numBoneDisplayName; displayIdx++)
 				{
-					auto& display = pmdFile->m_boneDisplayLists[displayIdx];
-					Read(&display.m_englishNameExt, file);
+					auto& [m_name, m_displayList, m_englishNameExt] = pmdFile->m_boneDisplayLists[displayIdx];
+					Read(&m_englishNameExt, file);
 				}
 			}
 
@@ -379,24 +379,24 @@ namespace saba
 			}
 
 			pmdFile->m_rigidBodies.resize(rigidBodyCount);
-			for (auto& rb : pmdFile->m_rigidBodies)
+			for (auto& [m_rigidBodyName, m_boneIndex, m_groupIndex, m_groupTarget, m_shapeType, m_shapeWidth, m_shapeHeight, m_shapeDepth, m_pos, m_rot, m_rigidBodyWeight, m_rigidBodyPosDimmer, m_rigidBodyRotDimmer, m_rigidBodyRecoil, m_rigidBodyFriction, m_rigidBodyType] : pmdFile->m_rigidBodies)
 			{
-				Read(&rb.m_rigidBodyName, file);
-				Read(&rb.m_boneIndex, file);
-				Read(&rb.m_groupIndex, file);
-				Read(&rb.m_groupTarget, file);
-				Read(&rb.m_shapeType, file);
-				Read(&rb.m_shapeWidth, file);
-				Read(&rb.m_shapeHeight, file);
-				Read(&rb.m_shapeDepth, file);
-				Read(&rb.m_pos, file);
-				Read(&rb.m_rot, file);
-				Read(&rb.m_rigidBodyWeight, file);
-				Read(&rb.m_rigidBodyPosDimmer, file);
-				Read(&rb.m_rigidBodyRotDimmer, file);
-				Read(&rb.m_rigidBodyRecoil, file);
-				Read(&rb.m_rigidBodyFriction, file);
-				Read(&rb.m_rigidBodyType, file);
+				Read(&m_rigidBodyName, file);
+				Read(&m_boneIndex, file);
+				Read(&m_groupIndex, file);
+				Read(&m_groupTarget, file);
+				Read(&m_shapeType, file);
+				Read(&m_shapeWidth, file);
+				Read(&m_shapeHeight, file);
+				Read(&m_shapeDepth, file);
+				Read(&m_pos, file);
+				Read(&m_rot, file);
+				Read(&m_rigidBodyWeight, file);
+				Read(&m_rigidBodyPosDimmer, file);
+				Read(&m_rigidBodyRotDimmer, file);
+				Read(&m_rigidBodyRecoil, file);
+				Read(&m_rigidBodyFriction, file);
+				Read(&m_rigidBodyType, file);
 			}
 
 			return !file.IsBad();
@@ -416,19 +416,19 @@ namespace saba
 			}
 
 			pmdFile->m_joints.resize(jointCount);
-			for (auto& joint : pmdFile->m_joints)
+			for (auto& [m_jointName, m_rigidBodyA, m_rigidBodyB, m_jointPos, m_jointRot, m_constrainPos1, m_constrainPos2, m_constrainRot1, m_constrainRot2, m_springPos, m_springRot] : pmdFile->m_joints)
 			{
-				Read(&joint.m_jointName, file);
-				Read(&joint.m_rigidBodyA, file);
-				Read(&joint.m_rigidBodyB, file);
-				Read(&joint.m_jointPos, file);
-				Read(&joint.m_jointRot, file);
-				Read(&joint.m_constrainPos1, file);
-				Read(&joint.m_constrainPos2, file);
-				Read(&joint.m_constrainRot1, file);
-				Read(&joint.m_constrainRot2, file);
-				Read(&joint.m_springPos, file);
-				Read(&joint.m_springRot, file);
+				Read(&m_jointName, file);
+				Read(&m_rigidBodyA, file);
+				Read(&m_rigidBodyB, file);
+				Read(&m_jointPos, file);
+				Read(&m_jointRot, file);
+				Read(&m_constrainPos1, file);
+				Read(&m_constrainPos2, file);
+				Read(&m_constrainRot1, file);
+				Read(&m_constrainRot2, file);
+				Read(&m_springPos, file);
+				Read(&m_springRot, file);
 			}
 
 			return !file.IsBad();
